Beam Drag and Rotation + Cursor + Cinemachine 3
Re-ordered the Beams in our UI deck. Added Beam Drag and high-lighted when beam is picked up, 'A' rotates the Beam counterclockwise and 'D' clockwise. Added Cinemachine package. For some reason I can't use the Cinemachine directive... Something fishy is going on... Also added a cute cursor because why not :)